add existing vault command
